% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bino_reg.R
\name{beta2der}
\alias{beta2der}
\title{Second Derivative of the Log-Likelihood (Negative Observed Information)}
\usage{
beta2der(X, beta.t, n)
}
\arguments{
\item{X}{A matrix of covariates}

\item{beta.t}{A numeric vector of current coefficient estimates}

\item{n}{A numeric vector of binomial denominators (trials)}
}
\value{
A numeric matrix representing the negative observed information
}
\description{
Computes the Hessian matrix for binomial regression.
}
\examples{
data(mtcars)
X <- model.matrix(~ hp + wt + cyl, data = mtcars)
n <- rep(1, nrow(mtcars))  # Bernoulli trials
beta_init <- rep(0, ncol(X))

# Compute Hessian
hessian <- beta2der(X, beta_init, n)
eigen(hessian)$values  # Should all be negative
}
